REFACTOR: BookmarkForm 생성 전담 Factory로 변경
